BREAKING(io): remove `MultiReader`
