+++ io/mod.ts
@@ -22,7 +22,6 @@ export * from "./buffer.ts";
 export * from "./copy.ts";
 export * from "./iterate_reader.ts";
 export * from "./limited_reader.ts";
-export * from "./multi_reader.ts";
 export * from "./read_all.ts";
 export * from "./read_delim.ts";
 export * from "./reader_from_stream_reader.ts";
