--- io/multi_reader_test.ts
@@ -1,14 +1,0 @@
-// Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
-
-import { assertEquals } from "@std/assert";
-import { MultiReader } from "./multi_reader.ts";
-import { StringWriter } from "./string_writer.ts";
-import { copy } from "./copy.ts";
-import { StringReader } from "./string_reader.ts";
-
-Deno.test("ioMultiReader", async function () {
-  const r = new MultiReader([new StringReader("abc"), new StringReader("def")]);
-  const w = new StringWriter();
-  await copy(r, w);
-  assertEquals(w.toString(), "abcdef");
-});
